requirements.txt
faust-streaming==0.11.3
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.37
pymysql==1.1.1
confluent-kafka==2.8.0